+++ lib/client/client.dart
@@ -49,6 +49,7 @@ import "../commands/fortnite/stw/daily.dart";
 import "../commands/fortnite/stw/dupe.dart";
 import "../commands/fortnite/stw/hero_loadout.dart";
 import "../commands/fortnite/stw/xpboost/xpboost.dart";
+import "../commands/fortnite/stw/inventory/inventory.dart";
 
 typedef NullableString = String;
 typedef NullableIUser = IUser;
@@ -158,6 +159,7 @@ class Client {
     _commands.addCommand(dupeCommand);
     _commands.addCommand(stwHeroLoadoutCommand);
     _commands.addCommand(xpBoostCommand);
+    _commands.addCommand(inventoryCommand);
 
     /// handle commands error
     handleCommandsError(this, _commands);