+++ lib/commands/fortnite/buy/buy.dart
@@ -1,0 +1,12 @@
+import "package:nyxx_commands/nyxx_commands.dart";
+
+import "br.dart";
+
+final ChatGroup buyCommand = ChatGroup(
+  "buy",
+  "Buy from fortnite item shop.",
+  children: [
+    brBuyCommand,
+  ],
+  checks: [],
+);