+++ lib/commands/fortnite/shop/br.dart
@@ -1,5 +1,6 @@
 import "package:nyxx/nyxx.dart";
 import "package:nyxx_commands/nyxx_commands.dart";
+import "../../../fishstick_dart.dart";
 
 final ChatCommand brShopCommand = ChatCommand(
   "br",
@@ -8,7 +9,7 @@ final ChatCommand brShopCommand = ChatCommand(
     "br_shop_command",
     (IContext ctx) async {
       await ctx.respond(MessageBuilder.content(
-          "https://fishstickbot.com/api/shop.png?v=${DateTime.now().millisecondsSinceEpoch}"));
+          "https://fishstickbot.com/api/shop.png?v=${client.systemJobs.catalogManagerSystemJob.brCatalog.uid}"));
     },
   ),
 );